feat: create reusable device button component
